requirements_poc.txt
Django==3.1.1
mysqlclient==2.0.1
djangorestframework == 3.11.1
pylint == 2.6.0